LPS-123900 Rename js to external video
