+++ modules/apps/document-library/document-library-external-video/src/main/resources/META-INF/resources/js/components/ExternalVideoPreview.js
